pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["al-ov71 <klever1987@yandex.ru>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
pytest-cov = "^4.1.0"
pyyaml = "^6.0.1"
requests = "^2.31.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
